# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PPS Business Cards API",
        "description": "API documentation for PPS Business Cards",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local server"
        },
        {
            "url": "https://test.ppsbusinesscards.de/",
            "description": "Test server"
        },
        {
            "url": "https://app.ppsbusinesscards.de/",
            "description": "Production server"
        }
    ],
    "paths": {
...
        "/api/v1/nfc/multi-assign": {
            "post": {
                "tags": [
                    "NFC Cards"
                ],
                "summary": "Assign multiple NFC cards to multiple employees",
                "description": "Assigns NFC cards to employees. All IDs must belong to the company from the API token.",
                "operationId": "ddfaa7d09bc1fb3cd00a510f71af798c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "assignments": {
                                        "description": "Array of assignments",
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "employee",
                                                "nfc_ids"
                                            ],
                                            "properties": {
                                                "employee": {
                                                    "description": "Employee ID to assign NFC cards to",
                                                    "type": "integer",
                                                    "example": 25
                                                },
                                                "nfc_ids": {
                                                    "description": "Array of NFC card IDs to assign",
                                                    "type": "array",
                                                    "items": {
                                                        "type": "integer",
                                                        "example": 101
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful assignment",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "2 assignment(s) completed successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "assignments": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "employee_id": {
                                                                "type": "integer",
                                                                "example": 25
                                                            },
                                                            "assigned_nfc_ids": {
                                                                "type": "array",
                                                                "items": {
                                                                    "type": "integer",
                                                                    "example": 101
                                                                }
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized / Invalid company ownership"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/nfc/multi-unassign": {
            "post": {
                "tags": [
                    "NFC Cards"
                ],
                "summary": "Unassign multiple NFC cards from employees",
                "description": "Unassigns NFC cards. All IDs must belong to the company from the API token.",
                "operationId": "ea9755059211eb073654c506a50f69a6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nfc_ids"
                                ],
                                "properties": {
                                    "nfc_ids": {
                                        "description": "Array of NFC card IDs to unassign",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 101
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful unassignment",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "5 NFC card(s) unassigned successfully."
                                        },
                                        "data": {
                                            "properties": {
                                                "unassigned_nfc_ids": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "integer",
                                                        "example": 101
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized / Invalid company ownership"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the format **Bearer &lt;token>**",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}